Custom doze: use ExecutorService to handle sensor listener
maybe this is useless, but out of curiosity let's try what Google
did for systemui here:
ezio84/abc_frameworks_base@fabc743

registering a sensor seems to be an expensive operation, and we do it on each
screen-on event, so moving it to an asynchronous task looks like
a good idea anyway.

The interesting part about ExecutorService api:
By
moving all non-essential binder calls of the main thread or to the
next frame, we bring this down to 5ms, such that window animation
and Keyguard animation starts about at the same time.
https://developer.android.com/reference/java/util/concurrent/ExecutorService
Change-Id: Id33a200b68e2f313af4258c6109c7b80aa93329c